Reupload due to lost logs + ball movement
Reupload of progress:
1. Screen is displaying and says "PING PONG"
Game Border and score display have been created.
2. Player and Enemy Stick now exist, Player stick also moves with Up and Down arrows.
3. The ball has been added.

New update:
The ball now collides with the roof and floor and will bounce off each in the opposite direction
Also if the ball gets past a stick the opposing stick gets 1 point and the ball is reset ready for the next game.